commit json writer/reader for one object

+++ Sources/DataServer/Modules/Category/src/main/java/com/rlrg/dataserver/category/repository/CategoryRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.rlrg.dataserver.category.entity.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long>, JpaSpecificationExecutor<Category>{
+public interface CategoryRepository extends JpaRepository<Category, Integer>, JpaSpecificationExecutor<Category>{
 	
 	@Query("Select c FROM Category WHERE c.status = :status")
 	public List<Category> getAllCategoryByStatus(boolean status);
